Refactor printf implementation for better handling of negative values and format specifiers; update linker script to include all text sections

+++ kernel/src/stdlib/numbers/atoi.c
@@ -13,5 +13,5 @@ int atoi(const char *str) {
         str++;
     }
     nbr = strtonbr(str);
-    return (sign * nbr);
+    return (int)((long long)sign * (long long)nbr);
 }
